+++ app/src/main/kotlin/ac/mdiq/podcini/net/download/EpisodeAdrDLManager.kt
@@ -9,7 +9,6 @@ import ac.mdiq.podcini.net.download.EpisodeAdrDLManager.Companion.WORK_DATA_PROG
 import ac.mdiq.podcini.net.download.EpisodeDLManager.Companion.updateDB
 import ac.mdiq.podcini.net.utils.NetworkUtils.mobileAllowEpisodeDownload
 import ac.mdiq.podcini.playback.base.InTheatre.actQueue
-import ac.mdiq.podcini.storage.database.addDownloadStatus
 import ac.mdiq.podcini.storage.database.addToAssQueue
 import ac.mdiq.podcini.storage.database.appAttribs
 import ac.mdiq.podcini.storage.database.appPrefs
@@ -19,6 +18,7 @@ import ac.mdiq.podcini.storage.database.removeFromAllQueues
 import ac.mdiq.podcini.storage.database.removeFromQueue
 import ac.mdiq.podcini.storage.database.upsert
 import ac.mdiq.podcini.storage.database.upsertBlk
+import ac.mdiq.podcini.storage.model.DownloadResult.Companion.addDownloadStatus
 import ac.mdiq.podcini.storage.model.Episode
 import ac.mdiq.podcini.storage.specs.EpisodeState
 import ac.mdiq.podcini.storage.utils.quietlyDeleteFile