+++ app/src/main/kotlin/ac/mdiq/podcini/storage/model/DownloadResult.kt
@@ -3,15 +3,15 @@ package ac.mdiq.podcini.storage.model
 import ac.mdiq.podcini.net.download.DownloadError
 import ac.mdiq.podcini.net.download.DownloadError.Companion.fromCode
 import ac.mdiq.podcini.net.download.RequestTye
+import ac.mdiq.podcini.storage.database.realm
+import ac.mdiq.podcini.storage.database.runOnIOScope
+import ac.mdiq.podcini.storage.database.upsert
+import ac.mdiq.podcini.storage.utils.nowInMillis
+import ac.mdiq.podcini.utils.Logd
 import io.github.xilinjia.krdb.types.RealmObject
 import io.github.xilinjia.krdb.types.annotations.Ignore
 import io.github.xilinjia.krdb.types.annotations.PrimaryKey
-import ac.mdiq.podcini.storage.utils.nowInMillis
 
-
-/**
- * Contains status attributes for one download
- */
 class DownloadResult : RealmObject {
     /**
      * A human-readable string which is shown to the user so that he can
@@ -36,11 +36,6 @@ class DownloadResult : RealmObject {
         }
     var reasonCode: Int = 0
 
-//    @Ignore private var completionDate: Date = Date()
-//        set(value) {
-//            field = value
-//            completionTime = field.time
-//        }
     var completionTime: Long = 0L
     /**
      * A message which can be presented to the user to give more information.
@@ -52,21 +47,15 @@ class DownloadResult : RealmObject {
     var id: Long = 0L
         private set
 
-    constructor(title: String, feedfileId: Long, feedfileType: Int, isSuccessful: Boolean, reason: DownloadError?, completionDate: Long, reasonDetailed: String) {
+    constructor(feedId: Long, title: String, reason: DownloadError?, successful: Boolean, reasonDetailed: String, feedfileType: Int = RequestTye.FEED.ordinal, completionDate: Long = nowInMillis()) {
         this.title = title
-        this.feedfileId = feedfileId
-        this.isSuccessful = isSuccessful
+        this.feedfileId = feedId
+        this.isSuccessful = successful
         this.feedfileType = feedfileType
         this.reason = reason
         this.completionTime = completionDate
         this.reasonDetailed = reasonDetailed
     }
-
-    /**
-     * Constructor for creating new completed downloads.
-     */
-    constructor(feedId: Long, title: String, reason: DownloadError, successful: Boolean, reasonDetailed: String)
-            : this(title, feedId, RequestTye.FEEDMEDIA.ordinal, successful, reason, nowInMillis(), reasonDetailed)
 
     constructor() : this(0L, "", DownloadError.ERROR_NOT_FOUND, false, "") {}
 
@@ -96,12 +85,27 @@ class DownloadResult : RealmObject {
     }
 
     companion object {
+        private const val TAG = "DownloadResult"
         /**
-         * Downloaders should use this constant for the size attribute if necessary
-         * so that the listadapters etc. can react properly.
+         * Downloaders should use this constant for the size attribute if necessary so that the listadapters etc. can react properly.
          */
         const val SIZE_UNKNOWN: Int = -1
 
         var idCounter: Long = -1
+
+        fun getFeedDownloadLog(feedId: Long): List<DownloadResult> {
+            Logd(TAG, "getFeedDownloadLog() called with: $feedId")
+            val dlog = realm.query(DownloadResult::class).query("feedfileId == $0", feedId).find().toMutableList()
+            dlog.sortWith { lhs, rhs ->  (rhs.completionTime - lhs.completionTime).toInt() }
+            return realm.copyFromRealm(dlog)
+        }
+
+        fun addDownloadStatus(status: DownloadResult) {
+            Logd(TAG, "addDownloadStatus called")
+            runOnIOScope {
+                if (status.id == 0L) status.setId()
+                upsert(status) {}
+            }
+        }
     }
 }